back_end/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==4.3.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
cryptography==44.0.2
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==5.0.1
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via -r requirements.in
pymysql==1.1.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   dotenv
requests==2.32.3
    # via -r requirements.in
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
